salesforce-hybrid-sim/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
neo4j>=5.28.0

